Extract common types
